Added messages popup to create/update item/itempacks. Made the material_cost_per_item form field required. Tweaked the messages.js to work with existing messages and not just ajax-created ones. Removed ItemPack page link from sidebar as those are managed via Item now.

+++ web/apps/market/views/item.py
@@ -1,4 +1,5 @@
 from django import urls
+from django.contrib import messages
 from django.views import generic
 
 from market import models as mkt_models, forms as mkt_forms
@@ -15,10 +16,16 @@ class ItemPackQuantitiesMixin:
         existing = set(self.object.item_packs.all().values_list('quantity', flat=True))
         remove_quantities = existing.difference(ipq)
         new_quantities = set(ipq).difference(existing)
-        self.object.item_packs.filter(quantity__in=remove_quantities).delete()
-        for new_qty in new_quantities:
-            self.object.item_packs.create(
-                quantity=new_qty, material_cost_per_pack=new_qty * self.object.material_cost_per_item)
+        if remove_quantities:
+            self.object.item_packs.filter(quantity__in=remove_quantities).delete()
+            remove_quantities_str = ", ".join(map(str, sorted(list(remove_quantities))))
+            messages.success(self.request, f"Removed packs {remove_quantities_str}")
+        if new_quantities:
+            for new_qty in new_quantities:
+                self.object.item_packs.create(
+                    quantity=new_qty, material_cost_per_pack=new_qty * self.object.material_cost_per_item)
+            new_quantities_str = ", ".join(map(str, sorted(list(new_quantities))))
+            messages.success(self.request, f"Added packs {new_quantities_str}")
         return ret
 
     def get_initial(self):
@@ -32,12 +39,22 @@ class ItemPackQuantitiesMixin:
 class ItemCreateView(ItemPackQuantitiesMixin, generic.CreateView):
     model = mkt_models.Item
 
+    def form_valid(self, form):
+        ret = super().form_valid(form)
+        messages.success(self.request, f"Created {self.object}.")
+        return ret
+
     def get_success_url(self):
         return urls.reverse('market:item_detail', args=(self.object.id,))
 
 
 class ItemDetailView(generic.DetailView):
     queryset = mkt_models.Item.objects.all()
+
+    def get_context_data(self, **kwargs):
+        context = super(ItemDetailView, self).get_context_data(**kwargs)
+        context['messages'] = messages.get_messages(self.request)
+        return context
 
 
 class ItemListView(generic.ListView):
@@ -46,6 +63,10 @@ class ItemListView(generic.ListView):
 
 class ItemUpdateView(ItemPackQuantitiesMixin, generic.UpdateView):
     model = mkt_models.Item
+
+    def form_valid(self, form):
+        messages.success(self.request, f"Updated {self.object}.")
+        return super().form_valid(form)
 
     def get_success_url(self):
         return urls.reverse('market:item_detail', args=(self.object.id,))
